Logging function now saves to "parser log.log"
The loggers never used to save to any location, only to the console. But
in working on the thread racing app I took some parts of the framework
away. I reinstated everything and changed to a fileHandler and had it
output to the file specified in the title.
Note that there's another logger that is declared at Player. There
doesn't seem to be any logging calls attached to it, but regardless it
saves to "logs.log"