Fixed the bug where scenarios with sound files played incorrectly
Simple fix, just had it create the test file in the same directory as
the loaded scenario if a scenario was ever loaded. This should prevent
that problem from happening regardless of where the loaded scenario is
from since it creates the temp file in the same directory. I chose not
to have it use the same file that was loaded because if I had, then
everytime you test the file would be overridden. This way, if you make
changes and are happy with the test and want to override the old file
you can then manually save over it after you test it.

New scenarios still function the same way and save the file in the same
location I had set before (user.dir)