Done with function counter
Now the authoring app tracks every use of the functions across different
instances of the JVM, it still prints out to the console as it did
before, however, it also prints to a file called functionCounter.txt.
This is also the file it saves the counters to so it can load them on
the next iteration.